requirements.txt
requests
beautifulsoup4
selenium
lxml
google-generativeai
webdriver_manager